Network.AWS.CloudTrail.AddTags

Description

Adds one or more tags to a trail, up to a limit of 50. Tags must be unique per trail. Overwrites an existing tag's value when a new value is specified for an existing tag key. If you specify a key without a value, the tag will be created with the specified key and a value of null. You can tag a trail that applies to all regions only from the region in which the trail was created (that is, from its home region).

Creating a Request

addTags

Arguments

:: Text
atResourceId

-> AddTags

Creates a value of AddTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atTagsList - Contains a list of CloudTrail tags, up to a limit of 50

atResourceId - Specifies the ARN of the trail to which one or more tags will be added. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

data AddTags

Specifies the tags to add to a trail.

See: addTags smart constructor.

Request Lenses

atTagsList :: Lens' AddTags [Tag]

Contains a list of CloudTrail tags, up to a limit of 50

atResourceId :: Lens' AddTags Text

Specifies the ARN of the trail to which one or more tags will be added. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

Destructuring the Response

addTagsResponse

Arguments

:: Int
atrsResponseStatus

-> AddTagsResponse

Creates a value of AddTagsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atrsResponseStatus - -- | The response status code.

data AddTagsResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: addTagsResponse smart constructor.

Response Lenses

atrsResponseStatus :: Lens' AddTagsResponse Int

- | The response status code.