Network.AWS.CloudTrail.CreateTrail

Description

From the command line, use 'create-subscription'.

Creates a trail that specifies the settings for delivery of log data to an Amazon S3 bucket.

http://docs.aws.amazon.com/awscloudtrail/latest/APIReference/API_CreateTrail.html

Request

data CreateTrail

Request constructor

createTrail

Arguments

:: Text
ctName

-> Text
ctS3BucketName

-> CreateTrail

CreateTrail constructor.

The fields accessible through corresponding lenses are:

ctCloudWatchLogsLogGroupArn :: Maybe Text

ctCloudWatchLogsRoleArn :: Maybe Text

ctIncludeGlobalServiceEvents :: Maybe Bool

ctName :: Text

ctS3BucketName :: Text

ctS3KeyPrefix :: Maybe Text

ctSnsTopicName :: Maybe Text

Request lenses

ctCloudWatchLogsLogGroupArn :: Lens' CreateTrail (Maybe Text)

Specifies a log group name using an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered. Not required unless you specify CloudWatchLogsRoleArn.

ctCloudWatchLogsRoleArn :: Lens' CreateTrail (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user’s log group.

ctIncludeGlobalServiceEvents :: Lens' CreateTrail (Maybe Bool)

Specifies whether the trail is publishing events from global services such as IAM to the log files.

ctName :: Lens' CreateTrail Text

Specifies the name of the trail.

ctS3BucketName :: Lens' CreateTrail Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

ctS3KeyPrefix :: Lens' CreateTrail (Maybe Text)

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

ctSnsTopicName :: Lens' CreateTrail (Maybe Text)

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

Response

data CreateTrailResponse

Response constructor

createTrailResponse :: CreateTrailResponse

CreateTrailResponse constructor.

The fields accessible through corresponding lenses are:

ctrCloudWatchLogsLogGroupArn :: Maybe Text

ctrCloudWatchLogsRoleArn :: Maybe Text

ctrIncludeGlobalServiceEvents :: Maybe Bool

ctrName :: Maybe Text

ctrS3BucketName :: Maybe Text

ctrS3KeyPrefix :: Maybe Text

ctrSnsTopicName :: Maybe Text

Response lenses

ctrCloudWatchLogsLogGroupArn :: Lens' CreateTrailResponse (Maybe Text)

Specifies the Amazon Resource Name (ARN) of the log group to which CloudTrail logs will be delivered.

ctrCloudWatchLogsRoleArn :: Lens' CreateTrailResponse (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user’s log group.

ctrIncludeGlobalServiceEvents :: Lens' CreateTrailResponse (Maybe Bool)

Specifies whether the trail is publishing events from global services such as IAM to the log files.

ctrName :: Lens' CreateTrailResponse (Maybe Text)

Specifies the name of the trail.

ctrS3BucketName :: Lens' CreateTrailResponse (Maybe Text)

Specifies the name of the Amazon S3 bucket designated for publishing log files.

ctrS3KeyPrefix :: Lens' CreateTrailResponse (Maybe Text)

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

ctrSnsTopicName :: Lens' CreateTrailResponse (Maybe Text)

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.