Network.AWS.CloudTrail.DeleteTrail

Description

Deletes a trail. This operation must be called from the region in which the trail was created. DeleteTrail cannot be called on the shadow trails (replicated trails in other regions) of a trail that is enabled in all regions.

Creating a Request

deleteTrail

Arguments

:: Text
dtName

-> DeleteTrail

Creates a value of DeleteTrail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtName - Specifies the name or the CloudTrail ARN of the trail to be deleted. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

data DeleteTrail

The request that specifies the name of a trail to delete.

See: deleteTrail smart constructor.

Request Lenses

dtName :: Lens' DeleteTrail Text

Specifies the name or the CloudTrail ARN of the trail to be deleted. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

Destructuring the Response

deleteTrailResponse

Arguments

:: Int
drsResponseStatus

-> DeleteTrailResponse

Creates a value of DeleteTrailResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsResponseStatus - -- | The response status code.

data DeleteTrailResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: deleteTrailResponse smart constructor.

Response Lenses

drsResponseStatus :: Lens' DeleteTrailResponse Int

- | The response status code.