Network.AWS.CloudTrail.GetEventSelectors

Description

Describes the settings for the event selectors that you configured for your trail. The information returned for your event selectors includes the following:

The S3 objects that you are logging for data events.

If your event selector includes management events.

If your event selector includes read-only events, write-only events, or all.

For more information, see Logging Data and Management Events for Trails in the AWS CloudTrail User Guide .

Creating a Request

getEventSelectors

Arguments

:: Text
gesTrailName

-> GetEventSelectors

Creates a value of GetEventSelectors with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gesTrailName - Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements: * Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-) * Start with a letter or number, and end with a letter or number * Be between 3 and 128 characters * Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are invalid. * Not be in IP address format (for example, 192.168.5.4) If you specify a trail ARN, it must be in the format: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

data GetEventSelectors

See: getEventSelectors smart constructor.

Request Lenses

gesTrailName :: Lens' GetEventSelectors Text

Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements: * Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-) * Start with a letter or number, and end with a letter or number * Be between 3 and 128 characters * Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are invalid. * Not be in IP address format (for example, 192.168.5.4) If you specify a trail ARN, it must be in the format: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

Destructuring the Response

getEventSelectorsResponse

Arguments

:: Int
gesrsResponseStatus

-> GetEventSelectorsResponse

Creates a value of GetEventSelectorsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gesrsTrailARN - The specified trail ARN that has the event selectors.

gesrsEventSelectors - The event selectors that are configured for the trail.

gesrsResponseStatus - -- | The response status code.

data GetEventSelectorsResponse

See: getEventSelectorsResponse smart constructor.

Response Lenses

gesrsTrailARN :: Lens' GetEventSelectorsResponse (Maybe Text)

The specified trail ARN that has the event selectors.

gesrsEventSelectors :: Lens' GetEventSelectorsResponse [EventSelector]

The event selectors that are configured for the trail.

gesrsResponseStatus :: Lens' GetEventSelectorsResponse Int

- | The response status code.