Network.AWS.CloudTrail.GetTrailStatus

Description

Returns a JSON-formatted list of information about the specified trail. Fields include information on delivery errors, Amazon SNS and Amazon S3 errors, and start and stop logging times for each trail.

http://docs.aws.amazon.com/awscloudtrail/latest/APIReference/API_GetTrailStatus.html

Request

data GetTrailStatus

Request constructor

getTrailStatus

Arguments

:: Text
gtsName

-> GetTrailStatus

GetTrailStatus constructor.

The fields accessible through corresponding lenses are:

gtsName :: Text

Request lenses

gtsName :: Lens' GetTrailStatus Text

The name of the trail for which you are requesting the current status.

Response

data GetTrailStatusResponse

Response constructor

getTrailStatusResponse :: GetTrailStatusResponse

GetTrailStatusResponse constructor.

The fields accessible through corresponding lenses are:

gtsrIsLogging :: Maybe Bool

gtsrLatestCloudWatchLogsDeliveryError :: Maybe Text

gtsrLatestCloudWatchLogsDeliveryTime :: Maybe UTCTime

gtsrLatestDeliveryError :: Maybe Text

gtsrLatestDeliveryTime :: Maybe UTCTime

gtsrLatestNotificationError :: Maybe Text

gtsrLatestNotificationTime :: Maybe UTCTime

gtsrStartLoggingTime :: Maybe UTCTime

gtsrStopLoggingTime :: Maybe UTCTime

Response lenses

gtsrIsLogging :: Lens' GetTrailStatusResponse (Maybe Bool)

Whether the CloudTrail is currently logging AWS API calls.

gtsrLatestCloudWatchLogsDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any CloudWatch Logs error that CloudTrail encountered when attempting to deliver logs to CloudWatch Logs.

gtsrLatestCloudWatchLogsDeliveryTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Displays the most recent date and time when CloudTrail delivered logs to CloudWatch Logs.

gtsrLatestDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver log files to the designated bucket. For more information see the topic Error Responses in the Amazon S3 API Reference.

gtsrLatestDeliveryTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the date and time that CloudTrail last delivered log files to an account's Amazon S3 bucket.

gtsrLatestNotificationError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon SNS error that CloudTrail encountered when attempting to send a notification. For more information about Amazon SNS errors, see the Amazon SNS Developer Guide.

gtsrLatestNotificationTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the date and time of the most recent Amazon SNS notification that CloudTrail has written a new log file to an account's Amazon S3 bucket.

gtsrStartLoggingTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the most recent date and time when CloudTrail started recording API calls for an AWS account.

gtsrStopLoggingTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the most recent date and time when CloudTrail stopped recording API calls for an AWS account.