Network.AWS.CloudTrail.ListPublicKeys

Description

Returns all public keys whose private keys were used to sign the digest files within the specified time range. The public key is needed to validate digest files that were signed with its corresponding private key.

Creating a Request

listPublicKeys :: ListPublicKeys

Creates a value of ListPublicKeys with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpkStartTime - Optionally specifies, in UTC, the start of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used, and the current public key is returned.

lpkNextToken - Reserved for future use.

lpkEndTime - Optionally specifies, in UTC, the end of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used.

data ListPublicKeys

Requests the public keys for a specified time range.

See: listPublicKeys smart constructor.

Request Lenses

lpkStartTime :: Lens' ListPublicKeys (Maybe UTCTime)

Optionally specifies, in UTC, the start of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used, and the current public key is returned.

lpkNextToken :: Lens' ListPublicKeys (Maybe Text)

Reserved for future use.

lpkEndTime :: Lens' ListPublicKeys (Maybe UTCTime)

Optionally specifies, in UTC, the end of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used.

Destructuring the Response

listPublicKeysResponse

Arguments

:: Int
lpkrsResponseStatus

-> ListPublicKeysResponse

Creates a value of ListPublicKeysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpkrsPublicKeyList - Contains an array of PublicKey objects.

lpkrsNextToken - Reserved for future use.

lpkrsResponseStatus - -- | The response status code.

data ListPublicKeysResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: listPublicKeysResponse smart constructor.

Response Lenses

lpkrsPublicKeyList :: Lens' ListPublicKeysResponse [PublicKey]

Contains an array of PublicKey objects.

lpkrsNextToken :: Lens' ListPublicKeysResponse (Maybe Text)

Reserved for future use.

lpkrsResponseStatus :: Lens' ListPublicKeysResponse Int

- | The response status code.