Network.AWS.CloudTrail.ListTags

Description

Lists the tags for the trail in the current region.

Creating a Request

listTags :: ListTags

Creates a value of ListTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltNextToken - Reserved for future use.

ltResourceIdList - Specifies a list of trail ARNs whose tags will be listed. The list has a limit of 20 ARNs. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

data ListTags

Specifies a list of trail tags to return.

See: listTags smart constructor.

Request Lenses

ltNextToken :: Lens' ListTags (Maybe Text)

Reserved for future use.

ltResourceIdList :: Lens' ListTags [Text]

Specifies a list of trail ARNs whose tags will be listed. The list has a limit of 20 ARNs. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

Destructuring the Response

listTagsResponse

Arguments

:: Int
ltrsResponseStatus

-> ListTagsResponse

Creates a value of ListTagsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltrsNextToken - Reserved for future use.

ltrsResourceTagList - A list of resource tags.

ltrsResponseStatus - -- | The response status code.

data ListTagsResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: listTagsResponse smart constructor.

Response Lenses

ltrsNextToken :: Lens' ListTagsResponse (Maybe Text)

Reserved for future use.

ltrsResourceTagList :: Lens' ListTagsResponse [ResourceTag]

A list of resource tags.

ltrsResponseStatus :: Lens' ListTagsResponse Int

- | The response status code.