Network.AWS.CloudTrail.LookupEvents

Description

Looks up API activity events captured by CloudTrail that create, update, or delete resources in your account. Events for a region can be looked up for the times in which you had CloudTrail turned on in that region during the last seven days. Lookup supports five different attributes: time range (defined by a start time and end time), user name, event name, resource type, and resource name. All attributes are optional. The maximum number of attributes that can be specified in any one lookup request are time range and one other attribute. The default number of results returned is 10, with a maximum of 50 possible. The response includes a token that you can use to get the next page of results. The rate of lookup requests is limited to one per second per account.

Events that occurred during the selected time range will not be available for lookup if CloudTrail logging was not enabled when the events occurred.

http://docs.aws.amazon.com/awscloudtrail/latest/APIReference/API_LookupEvents.html

Request

data LookupEvents

Request constructor

lookupEvents :: LookupEvents

LookupEvents constructor.

The fields accessible through corresponding lenses are:

leEndTime :: Maybe UTCTime

leLookupAttributes :: [LookupAttribute]

leMaxResults :: Maybe Natural

leNextToken :: Maybe Text

leStartTime :: Maybe UTCTime

Request lenses

leEndTime :: Lens' LookupEvents (Maybe UTCTime)

Specifies that only events that occur before or at the specified time are returned. If the specified end time is before the specified start time, an error is returned.

leLookupAttributes :: Lens' LookupEvents [LookupAttribute]

Contains a list of lookup attributes. Currently the list can contain only one item.

leMaxResults :: Lens' LookupEvents (Maybe Natural)

The number of events to return. Possible values are 1 through 50. The default is 10.

leNextToken :: Lens' LookupEvents (Maybe Text)

The token to use to get the next page of results after a previous API call. This token must be passed in with the same parameters that were specified in the the original call. For example, if the original call specified an AttributeKey of Username with a value of root, the call with NextToken should include those same parameters.

leStartTime :: Lens' LookupEvents (Maybe UTCTime)

Specifies that only events that occur after or at the specified time are returned. If the specified start time is after the specified end time, an error is returned.

Response

data LookupEventsResponse

Response constructor

lookupEventsResponse :: LookupEventsResponse

LookupEventsResponse constructor.

The fields accessible through corresponding lenses are:

lerEvents :: [Event]

lerNextToken :: Maybe Text

Response lenses

lerEvents :: Lens' LookupEventsResponse [Event]

A list of events returned based on the lookup attributes specified and the CloudTrail event. The events list is sorted by time. The most recent event is listed first.

lerNextToken :: Lens' LookupEventsResponse (Maybe Text)

The token to use to get the next page of results after a previous API call. If the token does not appear, there are no more results to return. The token must be passed in with the same parameters as the previous call. For example, if the original call specified an AttributeKey of Username with a value of root, the call with NextToken should include those same parameters.