Network.AWS.CloudTrail.PutEventSelectors

Description

Configures an event selector for your trail. Use event selectors to specify whether you want your trail to log management and/or data events. When an event occurs in your account, CloudTrail evaluates the event selectors in all trails. For each trail, if the event matches any event selector, the trail processes and logs the event. If the event doesn't match any event selector, the trail doesn't log the event.

Example

You create an event selector for a trail and specify that you want write-only events.

The EC2 GetConsoleOutput and RunInstances API operations occur in your account.

CloudTrail evaluates whether the events match your event selectors.

The RunInstances is a write-only event and it matches your event selector. The trail logs the event.

The GetConsoleOutput is a read-only event but it doesn't match your event selector. The trail doesn't log the event.

The PutEventSelectors operation must be called from the region in which the trail was created; otherwise, an InvalidHomeRegionException is thrown.

You can configure up to five event selectors for each trail. For more information, see Logging Data and Management Events for Trails in the AWS CloudTrail User Guide .

Creating a Request

putEventSelectors

Arguments

:: Text
pesTrailName

-> PutEventSelectors

Creates a value of PutEventSelectors with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pesTrailName - Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements: * Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-) * Start with a letter or number, and end with a letter or number * Be between 3 and 128 characters * Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are invalid. * Not be in IP address format (for example, 192.168.5.4) If you specify a trail ARN, it must be in the format: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

pesEventSelectors - Specifies the settings for your event selectors. You can configure up to five event selectors for a trail.

data PutEventSelectors

See: putEventSelectors smart constructor.

Request Lenses

pesTrailName :: Lens' PutEventSelectors Text

Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements: * Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-) * Start with a letter or number, and end with a letter or number * Be between 3 and 128 characters * Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are invalid. * Not be in IP address format (for example, 192.168.5.4) If you specify a trail ARN, it must be in the format: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

pesEventSelectors :: Lens' PutEventSelectors [EventSelector]

Specifies the settings for your event selectors. You can configure up to five event selectors for a trail.

Destructuring the Response

putEventSelectorsResponse

Arguments

:: Int
pesrsResponseStatus

-> PutEventSelectorsResponse

Creates a value of PutEventSelectorsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pesrsTrailARN - Specifies the ARN of the trail that was updated with event selectors. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

pesrsEventSelectors - Specifies the event selectors configured for your trail.

pesrsResponseStatus - -- | The response status code.

data PutEventSelectorsResponse

See: putEventSelectorsResponse smart constructor.

Response Lenses

pesrsTrailARN :: Lens' PutEventSelectorsResponse (Maybe Text)

Specifies the ARN of the trail that was updated with event selectors. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

pesrsEventSelectors :: Lens' PutEventSelectorsResponse [EventSelector]

Specifies the event selectors configured for your trail.

pesrsResponseStatus :: Lens' PutEventSelectorsResponse Int

- | The response status code.