Network.AWS.CloudTrail.Types

Description

CloudTrail is a web service that records AWS API calls for your AWS account and delivers log files to an Amazon S3 bucket. The recorded information includes the identity of the user, the start time of the AWS API call, the source IP address, the request parameters, and the response elements returned by the service.

Service

data CloudTrail

Supported version (2013-11-01) of the AWS CloudTrail service.

Errors

data CloudTrailError

A sum type representing possible errors returned by the CloudTrail service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

CloudTrailClient HttpException

CloudTrailSerializer String

CloudTrailService String

InsufficientS3BucketPolicyException
This exception is thrown when the policy on the S3 bucket is not sufficient.

InsufficientSnsTopicPolicyException
This exception is thrown when the policy on the SNS topic is not sufficient.

InvalidS3BucketNameException
This exception is thrown when the provided S3 bucket name is not valid.

InvalidS3PrefixException
This exception is thrown when the provided S3 prefix is not valid.

InvalidSnsTopicNameException
This exception is thrown when the provided SNS topic name is not valid.

InvalidTrailNameException
This exception is thrown when the provided trail name is not valid.

MaximumNumberOfTrailsExceededException
This exception is thrown when the maximum number of trails is reached.

S3BucketDoesNotExistException
This exception is thrown when the specified S3 bucket does not exist.

TrailAlreadyExistsException
This exception is thrown when the specified trail already exists.

TrailNotFoundException
This exception is thrown when the trail with the given name is not found.

_CloudTrailClient :: Prism' CloudTrailError HttpException

See: CloudTrailClient

_CloudTrailSerializer :: Prism' CloudTrailError String

See: CloudTrailSerializer

_CloudTrailService :: Prism' CloudTrailError String

See: CloudTrailService

_InsufficientS3BucketPolicyException :: Prism' CloudTrailError ()

This exception is thrown when the policy on the S3 bucket is not sufficient.

See: InsufficientS3BucketPolicyException

_InsufficientSnsTopicPolicyException :: Prism' CloudTrailError ()

This exception is thrown when the policy on the SNS topic is not sufficient.

See: InsufficientSnsTopicPolicyException

_InvalidS3BucketNameException :: Prism' CloudTrailError ()

This exception is thrown when the provided S3 bucket name is not valid.

See: InvalidS3BucketNameException

_InvalidS3PrefixException :: Prism' CloudTrailError ()

This exception is thrown when the provided S3 prefix is not valid.

See: InvalidS3PrefixException

_InvalidSnsTopicNameException :: Prism' CloudTrailError ()

This exception is thrown when the provided SNS topic name is not valid.

See: InvalidSnsTopicNameException

_InvalidTrailNameException :: Prism' CloudTrailError ()

This exception is thrown when the provided trail name is not valid.

See: InvalidTrailNameException

_MaximumNumberOfTrailsExceededException :: Prism' CloudTrailError ()

This exception is thrown when the maximum number of trails is reached.

See: MaximumNumberOfTrailsExceededException

_S3BucketDoesNotExistException :: Prism' CloudTrailError ()

This exception is thrown when the specified S3 bucket does not exist.

See: S3BucketDoesNotExistException

_TrailAlreadyExistsException :: Prism' CloudTrailError ()

This exception is thrown when the specified trail already exists.

See: TrailAlreadyExistsException

_TrailNotFoundException :: Prism' CloudTrailError ()

This exception is thrown when the trail with the given name is not found.

See: TrailNotFoundException

Trail

data Trail

The settings for a trail.

trail :: Trail

Smart constructor for the minimum required fields to construct a valid Trail data type.

Trail is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

S3BucketName :: Maybe Text

S3KeyPrefix :: Maybe Text

SnsTopicName :: Maybe Text

IncludeGlobalServiceEvents :: Maybe Bool

tName :: Lens' Trail (Maybe Text)

Name of the trail set by calling CreateTrail.

tS3BucketName :: Lens' Trail (Maybe Text)

Name of the Amazon S3 bucket into which CloudTrail delivers your trail files.

tS3KeyPrefix :: Lens' Trail (Maybe Text)

Value of the Amazon S3 prefix.

tSnsTopicName :: Lens' Trail (Maybe Text)

Name of the existing Amazon SNS topic that CloudTrail uses to notify the account owner when new CloudTrail log files have been delivered.

tIncludeGlobalServiceEvents :: Lens' Trail (Maybe Bool)

Set to True to include AWS API calls from AWS global services such as IAM. Otherwise, False.