Network.AWS.CloudTrail.UpdateTrail

Description

From the command line, use 'update-subscription'.

Updates the settings that specify delivery of log files. Changes to a trail do not require stopping the CloudTrail service. Use this action to designate an existing bucket for log delivery. If the existing bucket has previously been a target for CloudTrail log files, an IAM policy exists for the bucket.

http://docs.aws.amazon.com/awscloudtrail/latest/APIReference/API_UpdateTrail.html

Request

data UpdateTrail

Request constructor

updateTrail

Arguments

:: Text
utName

-> UpdateTrail

UpdateTrail constructor.

The fields accessible through corresponding lenses are:

utCloudWatchLogsLogGroupArn :: Maybe Text

utCloudWatchLogsRoleArn :: Maybe Text

utIncludeGlobalServiceEvents :: Maybe Bool

utName :: Text

utS3BucketName :: Maybe Text

utS3KeyPrefix :: Maybe Text

utSnsTopicName :: Maybe Text

Request lenses

utCloudWatchLogsLogGroupArn :: Lens' UpdateTrail (Maybe Text)

Specifies a log group name using an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered. Not required unless you specify CloudWatchLogsRoleArn.

utCloudWatchLogsRoleArn :: Lens' UpdateTrail (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user’s log group.

utIncludeGlobalServiceEvents :: Lens' UpdateTrail (Maybe Bool)

Specifies whether the trail is publishing events from global services such as IAM to the log files.

utName :: Lens' UpdateTrail Text

Specifies the name of the trail.

utS3BucketName :: Lens' UpdateTrail (Maybe Text)

Specifies the name of the Amazon S3 bucket designated for publishing log files.

utS3KeyPrefix :: Lens' UpdateTrail (Maybe Text)

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

utSnsTopicName :: Lens' UpdateTrail (Maybe Text)

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

Response

data UpdateTrailResponse

Response constructor

updateTrailResponse :: UpdateTrailResponse

UpdateTrailResponse constructor.

The fields accessible through corresponding lenses are:

utrCloudWatchLogsLogGroupArn :: Maybe Text

utrCloudWatchLogsRoleArn :: Maybe Text

utrIncludeGlobalServiceEvents :: Maybe Bool

utrName :: Maybe Text

utrS3BucketName :: Maybe Text

utrS3KeyPrefix :: Maybe Text

utrSnsTopicName :: Maybe Text

Response lenses

utrCloudWatchLogsLogGroupArn :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the Amazon Resource Name (ARN) of the log group to which CloudTrail logs will be delivered.

utrCloudWatchLogsRoleArn :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user’s log group.

utrIncludeGlobalServiceEvents :: Lens' UpdateTrailResponse (Maybe Bool)

Specifies whether the trail is publishing events from global services such as IAM to the log files.

utrName :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the name of the trail.

utrS3BucketName :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the name of the Amazon S3 bucket designated for publishing log files.

utrS3KeyPrefix :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

utrSnsTopicName :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.