Network.AWS.CloudTrail.V2013_11_01

Description

CloudTrail is a web service that records AWS API calls for your AWS account and delivers log files to an Amazon S3 bucket. The recorded information includes the identity of the user, the start time of the AWS API call, the source IP address, the request parameters, and the response elements returned by the service.

Documentation

module Network.AWS.CloudTrail.V2013_11_01.CreateTrail

module Network.AWS.CloudTrail.V2013_11_01.DeleteTrail

module Network.AWS.CloudTrail.V2013_11_01.DescribeTrails

module Network.AWS.CloudTrail.V2013_11_01.GetTrailStatus

module Network.AWS.CloudTrail.V2013_11_01.StartLogging

module Network.AWS.CloudTrail.V2013_11_01.StopLogging

module Network.AWS.CloudTrail.V2013_11_01.Types

module Network.AWS.CloudTrail.V2013_11_01.UpdateTrail