Network.AWS.CloudTrail.V2013_11_01.DescribeTrails

Description

Retrieves settings for the trail associated with the current region for your account.

Request

data DescribeTrails

Returns information about the trail.

Request constructor

mkDescribeTrails :: DescribeTrails

Smart constructor for the minimum required parameters to construct a valid DescribeTrails request.

Request lenses

dt1TrailNameList :: Lens' DescribeTrails [Text]

The trail returned.

Response

data DescribeTrailsResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

Response lenses

dtrrTrailList :: Lens' DescribeTrailsResponse [Trail]

The list of trails.