Network.AWS.CloudTrail.V2013_11_01.GetTrailStatus

Description

Returns a JSON-formatted list of information about the specified trail. Fields include information on delivery errors, Amazon SNS and Amazon S3 errors, and start and stop logging times for each trail.

Request

data GetTrailStatus

The name of a trail about which you want the current status.

Request constructor

mkGetTrailStatus

Arguments

:: Text
gtsName

-> GetTrailStatus

Smart constructor for the minimum required parameters to construct a valid GetTrailStatus request.

Request lenses

gtsName :: Lens' GetTrailStatus Text

The name of the trail for which you are requesting the current status.

Response

data GetTrailStatusResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

Response lenses

gtsrsIsLogging :: Lens' GetTrailStatusResponse (Maybe Bool)

Whether the CloudTrail is currently logging AWS API calls.

gtsrsLatestDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver log files to the designated bucket. For more information see the topic Error Responses in the Amazon S3 API Reference.

gtsrsLatestNotificationError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon SNS error that CloudTrail encountered when attempting to send a notification. For more information about Amazon SNS errors, see the Amazon SNS Developer Guide.

gtsrsLatestDeliveryTime :: Lens' GetTrailStatusResponse (Maybe ISO8601)

Specifies the date and time that CloudTrail last delivered log files to an account's Amazon S3 bucket.

gtsrsLatestNotificationTime :: Lens' GetTrailStatusResponse (Maybe ISO8601)

Specifies the date and time of the most recent Amazon SNS notification that CloudTrail has written a new log file to an account's Amazon S3 bucket.

gtsrsStartLoggingTime :: Lens' GetTrailStatusResponse (Maybe ISO8601)

Specifies the most recent date and time when CloudTrail started recording API calls for an AWS account.

gtsrsStopLoggingTime :: Lens' GetTrailStatusResponse (Maybe ISO8601)

Specifies the most recent date and time when CloudTrail stopped recording API calls for an AWS account.