Network.AWS.CloudTrail.V2013_11_01.Monadic

Description

CloudTrail is a web service that records AWS API calls for your AWS account and delivers log files to an Amazon S3 bucket. The recorded information includes the identity of the user, the start time of the AWS API call, the source IP address, the request parameters, and the response elements returned by the service.

This module is provided for convenience. It offers an alternative to the common idiom of supplying required fields to an operations's smart constructor, using the operation's lenses to modify additional fields, and then sending or paginating the request.

As an example: using Network.AWS.CloudTrail with the smart constructor and basic lens syntax, before explicitly calling send:

import Control.Monad.Trans.AWS
import Network.AWS.CloudTrail

send $ (mkOperationName w x)
     & onLensField1 .~ y
     & onLensField2 .~ z

Versus using Network.AWS.CloudTrail.Monadic with the State operator variants from Control.Lens.Setter such as .= to modify any additional request parameters before sending:

import Network.AWS.CloudTrail.Monadic

operationName w x $ do
    onLensField1 .= y
    onLensField2 .= z

CreateTrail

From the command line, use create-subscription. Creates a trail that specifies the settings for delivery of log data to an Amazon S3 bucket.

See: CreateTrail

createTrail

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ctName

-> Text
ctS3BucketName

-> State CreateTrail a

-> m CreateTrailResponse

createTrailCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ctName

-> Text
ctS3BucketName

-> State CreateTrail a

-> m (Either ServiceEr CreateTrailResponse)

DeleteTrail

Deletes a trail.

See: DeleteTrail

deleteTrail

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dtName

-> State DeleteTrail a

-> m DeleteTrailResponse

deleteTrailCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dtName

-> State DeleteTrail a

-> m (Either ServiceEr DeleteTrailResponse)

DescribeTrails

Retrieves settings for the trail associated with the current region for your account.

See: DescribeTrails

describeTrails :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeTrails a -> m DescribeTrailsResponse

describeTrailsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeTrails a -> m (Either ServiceEr DescribeTrailsResponse)

GetTrailStatus

Returns a JSON-formatted list of information about the specified trail. Fields include information on delivery errors, Amazon SNS and Amazon S3 errors, and start and stop logging times for each trail.

See: GetTrailStatus

getTrailStatus

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gtsName

-> State GetTrailStatus a

-> m GetTrailStatusResponse

getTrailStatusCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gtsName

-> State GetTrailStatus a

-> m (Either ServiceEr GetTrailStatusResponse)

StartLogging

Starts the recording of AWS API calls and log file delivery for a trail.

See: StartLogging

startLogging

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
slName

-> State StartLogging a

-> m StartLoggingResponse

startLoggingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
slName

-> State StartLogging a

-> m (Either ServiceEr StartLoggingResponse)

StopLogging

Suspends the recording of AWS API calls and log file delivery for the specified trail. Under most circumstances, there is no need to use this action. You can update a trail without stopping it first. This action is the only way to stop recording.

See: StopLogging

stopLogging

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
sl1Name

-> State StopLogging a

-> m StopLoggingResponse

stopLoggingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
sl1Name

-> State StopLogging a

-> m (Either ServiceEr StopLoggingResponse)

UpdateTrail

From the command line, use update-subscription. Updates the settings that specify delivery of log files. Changes to a trail do not require stopping the CloudTrail service. Use this action to designate an existing bucket for log delivery. If the existing bucket has previously been a target for CloudTrail log files, an IAM policy exists for the bucket.

See: UpdateTrail

updateTrail

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
utName

-> State UpdateTrail a

-> m UpdateTrailResponse

updateTrailCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
utName

-> State UpdateTrail a

-> m (Either ServiceEr UpdateTrailResponse)

Re-exported

module Network.AWS.CloudTrail.V2013_11_01

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()