Network.AWS.CloudTrail.V2013_11_01.StartLogging

Description

Starts the recording of AWS API calls and log file delivery for a trail.

Request

data StartLogging

The request to CloudTrail to start logging AWS API calls for an account.

Request constructor

mkStartLogging

Arguments

:: Text
slName

-> StartLogging

Smart constructor for the minimum required parameters to construct a valid StartLogging request.

The fields accessible through corresponding lenses are:

Name :: Text

Request lenses

slName :: Lens' StartLogging Text

The name of the trail for which CloudTrail logs AWS API calls.

Response

data StartLoggingResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

Response constructor

mkStartLoggingResponse :: StartLoggingResponse

Smart constructor for the minimum required parameters to construct a valid StartLoggingResponse response.

This constructor is provided for convenience and testing purposes.