Network.AWS.CloudTrail.V2013_11_01.StopLogging

Description

Suspends the recording of AWS API calls and log file delivery for the specified trail. Under most circumstances, there is no need to use this action. You can update a trail without stopping it first. This action is the only way to stop recording.

Documentation

data StopLogging

Constructors

StopLogging

Fields

_slrName :: Text

Communicates to CloudTrail the name of the trail for which to stop logging AWS API calls.

slrName :: Iso' StopLogging Text

data StopLoggingResponse

Constructors

StopLoggingResponse