Network.AWS.CloudTrail.V2013_11_01.Types

Description

CloudTrail is a web service that records AWS API calls for your AWS account and delivers log files to an Amazon S3 bucket. The recorded information includes the identity of the user, the start time of the AWS API call, the source IP address, the request parameters, and the response elements returned by the service.

Service

data CloudTrail

Supported version (2013-11-01) of the AWS CloudTrail service.

Trail

data Trail

The settings for a trail.

mkTrail :: Trail

Smart constructor for the minimum required fields to construct a valid Trail data type.

Trail is exclusively used in responses and this constructor is provided for convenience and testing purposes.

tName :: Lens' Trail (Maybe Text)

Name of the trail set by calling CreateTrail.

tS3BucketName :: Lens' Trail (Maybe Text)

Name of the Amazon S3 bucket into which CloudTrail delivers your trail files.

tS3KeyPrefix :: Lens' Trail (Maybe Text)

Value of the Amazon S3 prefix.

tSnsTopicName :: Lens' Trail (Maybe Text)

Name of the existing Amazon SNS topic that CloudTrail uses to notify the account owner when new CloudTrail log files have been delivered.

tIncludeGlobalServiceEvents :: Lens' Trail (Maybe Bool)

Set to True to include AWS API calls from AWS global services such as IAM. Otherwise, False.