Network.AWS.CloudTrail.V2013_11_01.UpdateTrail

Description

From the command line, use update-subscription. Updates the settings that specify delivery of log files. Changes to a trail do not require stopping the CloudTrail service. Use this action to designate an existing bucket for log delivery. If the existing bucket has previously been a target for CloudTrail log files, an IAM policy exists for the bucket.

Documentation

updateTrail

Arguments

:: Text
_utrName

-> UpdateTrail

Minimum specification for a UpdateTrail request.

data UpdateTrail

Constructors

UpdateTrail

Fields

_utrName :: Text

Specifies the name of the trail.

_utrIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_utrS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_utrSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

_utrS3BucketName :: Maybe Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

utrSnsTopicName :: Lens' UpdateTrail (Maybe Text)

utrS3KeyPrefix :: Lens' UpdateTrail (Maybe Text)

utrS3BucketName :: Lens' UpdateTrail (Maybe Text)

utrName :: Lens' UpdateTrail Text

utrIncludeGlobalServiceEvents :: Lens' UpdateTrail (Maybe Bool)

data UpdateTrailResponse

Constructors

UpdateTrailResponse

Fields

_utsIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_utsS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_utsSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

_utsName :: Maybe Text

Specifies the name of the trail.

_utsS3BucketName :: Maybe Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

utsSnsTopicName :: Lens' UpdateTrailResponse (Maybe Text)

utsS3KeyPrefix :: Lens' UpdateTrailResponse (Maybe Text)

utsS3BucketName :: Lens' UpdateTrailResponse (Maybe Text)

utsName :: Lens' UpdateTrailResponse (Maybe Text)

utsIncludeGlobalServiceEvents :: Lens' UpdateTrailResponse (Maybe Bool)