Network.AWS.CloudTrail.GetTrailStatus

Description

Returns a JSON-formatted list of information about the specified trail. Fields include information on delivery errors, Amazon SNS and Amazon S3 errors, and start and stop logging times for each trail. This operation returns trail status from a single region. To return trail status from all regions, you must call the operation on each region.

Creating a Request

getTrailStatus

Arguments

:: Text
gtsName

-> GetTrailStatus

Creates a value of GetTrailStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtsName - Specifies the name or the CloudTrail ARN of the trail for which you are requesting status. To get the status of a shadow trail (a replication of the trail in another region), you must specify its ARN. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

data GetTrailStatus

The name of a trail about which you want the current status.

See: getTrailStatus smart constructor.

Request Lenses

gtsName :: Lens' GetTrailStatus Text

Specifies the name or the CloudTrail ARN of the trail for which you are requesting status. To get the status of a shadow trail (a replication of the trail in another region), you must specify its ARN. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

Destructuring the Response

getTrailStatusResponse

Arguments

:: Int
gtsrsResponseStatus

-> GetTrailStatusResponse

Creates a value of GetTrailStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtsrsTimeLoggingStopped - This field is deprecated.

gtsrsLatestDeliveryError - Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver log files to the designated bucket. For more information see the topic Error Responses in the Amazon S3 API Reference.

gtsrsLatestDigestDeliveryTime - Specifies the date and time that CloudTrail last delivered a digest file to an account's Amazon S3 bucket.

gtsrsLatestNotificationAttemptSucceeded - This field is deprecated.

gtsrsStartLoggingTime - Specifies the most recent date and time when CloudTrail started recording API calls for an AWS account.

gtsrsLatestNotificationError - Displays any Amazon SNS error that CloudTrail encountered when attempting to send a notification. For more information about Amazon SNS errors, see the Amazon SNS Developer Guide .

gtsrsLatestDeliveryAttemptSucceeded - This field is deprecated.

gtsrsIsLogging - Whether the CloudTrail is currently logging AWS API calls.

gtsrsTimeLoggingStarted - This field is deprecated.

gtsrsLatestDigestDeliveryError - Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver a digest file to the designated bucket. For more information see the topic Error Responses in the Amazon S3 API Reference.

gtsrsLatestDeliveryAttemptTime - This field is deprecated.

gtsrsLatestDeliveryTime - Specifies the date and time that CloudTrail last delivered log files to an account's Amazon S3 bucket.

gtsrsLatestCloudWatchLogsDeliveryTime - Displays the most recent date and time when CloudTrail delivered logs to CloudWatch Logs.

gtsrsLatestCloudWatchLogsDeliveryError - Displays any CloudWatch Logs error that CloudTrail encountered when attempting to deliver logs to CloudWatch Logs.

gtsrsLatestNotificationTime - Specifies the date and time of the most recent Amazon SNS notification that CloudTrail has written a new log file to an account's Amazon S3 bucket.

gtsrsLatestNotificationAttemptTime - This field is deprecated.

gtsrsStopLoggingTime - Specifies the most recent date and time when CloudTrail stopped recording API calls for an AWS account.

gtsrsResponseStatus - -- | The response status code.

data GetTrailStatusResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: getTrailStatusResponse smart constructor.

Response Lenses

gtsrsTimeLoggingStopped :: Lens' GetTrailStatusResponse (Maybe Text)

This field is deprecated.

gtsrsLatestDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver log files to the designated bucket. For more information see the topic Error Responses in the Amazon S3 API Reference.

gtsrsLatestDigestDeliveryTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the date and time that CloudTrail last delivered a digest file to an account's Amazon S3 bucket.

gtsrsLatestNotificationAttemptSucceeded :: Lens' GetTrailStatusResponse (Maybe Text)

This field is deprecated.

gtsrsStartLoggingTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the most recent date and time when CloudTrail started recording API calls for an AWS account.

gtsrsLatestNotificationError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon SNS error that CloudTrail encountered when attempting to send a notification. For more information about Amazon SNS errors, see the Amazon SNS Developer Guide .

gtsrsLatestDeliveryAttemptSucceeded :: Lens' GetTrailStatusResponse (Maybe Text)

This field is deprecated.

gtsrsIsLogging :: Lens' GetTrailStatusResponse (Maybe Bool)

Whether the CloudTrail is currently logging AWS API calls.

gtsrsTimeLoggingStarted :: Lens' GetTrailStatusResponse (Maybe Text)

This field is deprecated.

gtsrsLatestDigestDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver a digest file to the designated bucket. For more information see the topic Error Responses in the Amazon S3 API Reference.

gtsrsLatestDeliveryAttemptTime :: Lens' GetTrailStatusResponse (Maybe Text)

This field is deprecated.

gtsrsLatestDeliveryTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the date and time that CloudTrail last delivered log files to an account's Amazon S3 bucket.

gtsrsLatestCloudWatchLogsDeliveryTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Displays the most recent date and time when CloudTrail delivered logs to CloudWatch Logs.

gtsrsLatestCloudWatchLogsDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any CloudWatch Logs error that CloudTrail encountered when attempting to deliver logs to CloudWatch Logs.

gtsrsLatestNotificationTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the date and time of the most recent Amazon SNS notification that CloudTrail has written a new log file to an account's Amazon S3 bucket.

gtsrsLatestNotificationAttemptTime :: Lens' GetTrailStatusResponse (Maybe Text)

This field is deprecated.

gtsrsStopLoggingTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the most recent date and time when CloudTrail stopped recording API calls for an AWS account.

gtsrsResponseStatus :: Lens' GetTrailStatusResponse Int

- | The response status code.