Network.AWS.CloudTrail.LookupEvents

Description

Looks up API activity events captured by CloudTrail that create, update, or delete resources in your account. Events for a region can be looked up for the times in which you had CloudTrail turned on in that region during the last seven days. Lookup supports the following attributes:

Event ID

Event name

Event source

Resource name

Resource type

User name

All attributes are optional. The default number of results returned is 10, with a maximum of 50 possible. The response includes a token that you can use to get the next page of results.

Important: The rate of lookup requests is limited to one per second per account. If this limit is exceeded, a throttling error occurs.

Important: Events that occurred during the selected time range will not be available for lookup if CloudTrail logging was not enabled when the events occurred.

This operation returns paginated results.

Creating a Request

lookupEvents :: LookupEvents

Creates a value of LookupEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

leStartTime - Specifies that only events that occur after or at the specified time are returned. If the specified start time is after the specified end time, an error is returned.

leLookupAttributes - Contains a list of lookup attributes. Currently the list can contain only one item.

leNextToken - The token to use to get the next page of results after a previous API call. This token must be passed in with the same parameters that were specified in the the original call. For example, if the original call specified an AttributeKey of Username with a value of root, the call with NextToken should include those same parameters.

leEndTime - Specifies that only events that occur before or at the specified time are returned. If the specified end time is before the specified start time, an error is returned.

leMaxResults - The number of events to return. Possible values are 1 through 50. The default is 10.

data LookupEvents

Contains a request for LookupEvents.

See: lookupEvents smart constructor.

Request Lenses

leStartTime :: Lens' LookupEvents (Maybe UTCTime)

Specifies that only events that occur after or at the specified time are returned. If the specified start time is after the specified end time, an error is returned.

leLookupAttributes :: Lens' LookupEvents [LookupAttribute]

Contains a list of lookup attributes. Currently the list can contain only one item.

leNextToken :: Lens' LookupEvents (Maybe Text)

The token to use to get the next page of results after a previous API call. This token must be passed in with the same parameters that were specified in the the original call. For example, if the original call specified an AttributeKey of Username with a value of root, the call with NextToken should include those same parameters.

leEndTime :: Lens' LookupEvents (Maybe UTCTime)

Specifies that only events that occur before or at the specified time are returned. If the specified end time is before the specified start time, an error is returned.

leMaxResults :: Lens' LookupEvents (Maybe Natural)

The number of events to return. Possible values are 1 through 50. The default is 10.

Destructuring the Response

lookupEventsResponse

Arguments

:: Int
lersResponseStatus

-> LookupEventsResponse

Creates a value of LookupEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lersNextToken - The token to use to get the next page of results after a previous API call. If the token does not appear, there are no more results to return. The token must be passed in with the same parameters as the previous call. For example, if the original call specified an AttributeKey of Username with a value of root, the call with NextToken should include those same parameters.

lersEvents - A list of events returned based on the lookup attributes specified and the CloudTrail event. The events list is sorted by time. The most recent event is listed first.

lersResponseStatus - -- | The response status code.

data LookupEventsResponse

Contains a response to a LookupEvents action.

See: lookupEventsResponse smart constructor.

Response Lenses

lersNextToken :: Lens' LookupEventsResponse (Maybe Text)

The token to use to get the next page of results after a previous API call. If the token does not appear, there are no more results to return. The token must be passed in with the same parameters as the previous call. For example, if the original call specified an AttributeKey of Username with a value of root, the call with NextToken should include those same parameters.

lersEvents :: Lens' LookupEventsResponse [Event]

A list of events returned based on the lookup attributes specified and the CloudTrail event. The events list is sorted by time. The most recent event is listed first.

lersResponseStatus :: Lens' LookupEventsResponse Int

- | The response status code.