Network.AWS.CloudTrail.Types

Description

Service Configuration

cloudTrail :: Service

API version 2013-11-01 of the Amazon CloudTrail SDK configuration.

Errors

_InvalidTimeRangeException :: AsError a => Getting (First ServiceError) a ServiceError

Occurs if the timestamp values are invalid. Either the start time occurs after the end time or the time range is outside the range of possible values.

_InsufficientS3BucketPolicyException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the policy on the S3 bucket is not sufficient.

_MaximumNumberOfTrailsExceededException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the maximum number of trails is reached.

_UnsupportedOperationException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the requested operation is not supported.

_KMSKeyDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is deprecated.

_InsufficientEncryptionPolicyException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the policy on the S3 bucket or KMS key is not sufficient.

_InsufficientSNSTopicPolicyException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the policy on the SNS topic is not sufficient.

_InvalidCloudWatchLogsRoleARNException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the provided role is not valid.

_TagsLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of tags per trail has exceeded the permitted amount. Currently, the limit is 50.

_CloudTrailARNInvalidException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when an operation is called with an invalid trail ARN. The format of a trail ARN is:

arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

_InvalidLookupAttributesException :: AsError a => Getting (First ServiceError) a ServiceError

Occurs when an invalid lookup attribute is specified.

_InvalidTrailNameException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the provided trail name is not valid. Trail names must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are invalid.

Not be in IP address format (for example, 192.168.5.4)

_InvalidSNSTopicNameException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the provided SNS topic name is not valid.

_ResourceTypeNotSupportedException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the specified resource type is not supported by CloudTrail.

_CloudWatchLogsDeliveryUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Cannot set a CloudWatch Logs delivery for this region.

_KMSKeyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the KMS key does not exist, or when the S3 bucket and the KMS key are not in the same region.

_TrailNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the trail with the given name is not found.

_InvalidEventSelectorsException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the PutEventSelectors operation is called with an invalid number of event selectors, data resources, or an invalid value for a parameter:

Specify a valid number of event selectors (1 to 5) for a trail.

Specify a valid number of data resources (1 to 250) for an event selector.

Specify a valid value for a parameter. For example, specifying the ReadWriteType parameter with a value of read-only is invalid.

_TrailNotProvidedException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is deprecated.

_InvalidS3BucketNameException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the provided S3 bucket name is not valid.

_InvalidCloudWatchLogsLogGroupARNException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the provided CloudWatch log group is not valid.

_KMSException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when there is an issue with the specified KMS key and the trail can’t be updated.

_S3BucketDoesNotExistException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the specified S3 bucket does not exist.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Invalid token or token that was previously used in a request with different parameters. This exception is thrown if the token is invalid.

_InvalidTagParameterException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the key or value specified for the tag does not match the regular expression ^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$@ .

_OperationNotPermittedException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the requested operation is not permitted.

_InvalidTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Reserved for future use.

_InvalidMaxResultsException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown if the limit specified is invalid.

_TrailAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the specified trail already exists.

_InvalidS3PrefixException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the provided S3 prefix is not valid.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the specified resource is not found.

_InvalidParameterCombinationException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the combination of parameters provided is not valid.

_InvalidKMSKeyIdException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the KMS key ARN is invalid.

_InvalidHomeRegionException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when an operation is called on a trail from a region other than the region in which the trail was created.

LookupAttributeKey

data LookupAttributeKey

Constructors

EventId

EventName

EventSource

ResourceName

ResourceType

Username

ReadWriteType

data ReadWriteType

Constructors

All

ReadOnly

WriteOnly

DataResource

data DataResource

The Amazon S3 objects that you specify in your event selectors for your trail to log data events. Data events are object-level API operations that access S3 objects, such as GetObject , DeleteObject , and PutObject . You can specify up to 250 S3 buckets and object prefixes for a trail.

Example

You create an event selector for a trail and specify an S3 bucket and an empty prefix, such as arn:aws:s3:::bucket-1/ .

You upload an image file to bucket-1 .

The PutObject API operation occurs on an object in the S3 bucket that you specified in the event selector. The trail processes and logs the event.

You upload another image file to a different S3 bucket named arn:aws:s3:::bucket-2 .

The event occurs on an object in an S3 bucket that you didn't specify in the event selector. The trail doesn’t log the event.

See: dataResource smart constructor.

dataResource :: DataResource

Creates a value of DataResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drValues - A list of ARN-like strings for the specified S3 objects. To log data events for all objects in an S3 bucket, specify the bucket and an empty object prefix such as arn:aws:s3:::bucket-1/ . The trail logs data events for all objects in this S3 bucket. To log data events for specific objects, specify the S3 bucket and object prefix such as arn:aws:s3:::bucket-1/example-images . The trail logs data events for objects in this S3 bucket that match the prefix.

drType - The resource type in which you want to log data events. You can specify only the following value: AWS::S3::Object .

drValues :: Lens' DataResource [Text]

A list of ARN-like strings for the specified S3 objects. To log data events for all objects in an S3 bucket, specify the bucket and an empty object prefix such as arn:aws:s3:::bucket-1/ . The trail logs data events for all objects in this S3 bucket. To log data events for specific objects, specify the S3 bucket and object prefix such as arn:aws:s3:::bucket-1/example-images . The trail logs data events for objects in this S3 bucket that match the prefix.

drType :: Lens' DataResource (Maybe Text)

The resource type in which you want to log data events. You can specify only the following value: AWS::S3::Object .

Event

data Event

Contains information about an event that was returned by a lookup request. The result includes a representation of a CloudTrail event.

See: event smart constructor.

event :: Event

Creates a value of Event with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eUsername - A user name or role name of the requester that called the API in the event returned.

eResources - A list of resources referenced by the event returned.

eEventTime - The date and time of the event returned.

eCloudTrailEvent - A JSON string that contains a representation of the event returned.

eEventName - The name of the event returned.

eEventSource - The AWS service that the request was made to.

eEventId - The CloudTrail ID of the event returned.

eUsername :: Lens' Event (Maybe Text)

A user name or role name of the requester that called the API in the event returned.

eResources :: Lens' Event [Resource]

A list of resources referenced by the event returned.

eEventTime :: Lens' Event (Maybe UTCTime)

The date and time of the event returned.

eCloudTrailEvent :: Lens' Event (Maybe Text)

A JSON string that contains a representation of the event returned.

eEventName :: Lens' Event (Maybe Text)

The name of the event returned.

eEventSource :: Lens' Event (Maybe Text)

The AWS service that the request was made to.

eEventId :: Lens' Event (Maybe Text)

The CloudTrail ID of the event returned.

EventSelector

data EventSelector

Use event selectors to specify whether you want your trail to log management and/or data events. When an event occurs in your account, CloudTrail evaluates the event selector for all trails. For each trail, if the event matches any event selector, the trail processes and logs the event. If the event doesn't match any event selector, the trail doesn't log the event.

You can configure up to five event selectors for a trail.

See: eventSelector smart constructor.

eventSelector :: EventSelector

Creates a value of EventSelector with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esDataResources - CloudTrail supports logging only data events for S3 objects. You can specify up to 250 S3 buckets and object prefixes for a trail. For more information, see Data Events in the AWS CloudTrail User Guide .

esReadWriteType - Specify if you want your trail to log read-only events, write-only events, or all. For example, the EC2 GetConsoleOutput is a read-only API operation and RunInstances is a write-only API operation. By default, the value is All .

esIncludeManagementEvents - Specify if you want your event selector to include management events for your trail. For more information, see Management Events in the AWS CloudTrail User Guide . By default, the value is true .

esDataResources :: Lens' EventSelector [DataResource]

CloudTrail supports logging only data events for S3 objects. You can specify up to 250 S3 buckets and object prefixes for a trail. For more information, see Data Events in the AWS CloudTrail User Guide .

esReadWriteType :: Lens' EventSelector (Maybe ReadWriteType)

Specify if you want your trail to log read-only events, write-only events, or all. For example, the EC2 GetConsoleOutput is a read-only API operation and RunInstances is a write-only API operation. By default, the value is All .

esIncludeManagementEvents :: Lens' EventSelector (Maybe Bool)

Specify if you want your event selector to include management events for your trail. For more information, see Management Events in the AWS CloudTrail User Guide . By default, the value is true .

LookupAttribute

data LookupAttribute

Specifies an attribute and value that filter the events returned.

See: lookupAttribute smart constructor.

lookupAttribute

Arguments

:: LookupAttributeKey
laAttributeKey

-> Text
laAttributeValue

-> LookupAttribute

Creates a value of LookupAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laAttributeKey - Specifies an attribute on which to filter the events returned.

laAttributeValue - Specifies a value for the specified AttributeKey.

laAttributeKey :: Lens' LookupAttribute LookupAttributeKey

Specifies an attribute on which to filter the events returned.

laAttributeValue :: Lens' LookupAttribute Text

Specifies a value for the specified AttributeKey.

PublicKey

data PublicKey

Contains information about a returned public key.

See: publicKey smart constructor.

publicKey :: PublicKey

Creates a value of PublicKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pkFingerprint - The fingerprint of the public key.

pkValidityEndTime - The ending time of validity of the public key.

pkValue - The DER encoded public key value in PKCS#1 format.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

pkValidityStartTime - The starting time of validity of the public key.

pkFingerprint :: Lens' PublicKey (Maybe Text)

The fingerprint of the public key.

pkValidityEndTime :: Lens' PublicKey (Maybe UTCTime)

The ending time of validity of the public key.

pkValue :: Lens' PublicKey (Maybe ByteString)

The DER encoded public key value in PKCS#1 format.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

pkValidityStartTime :: Lens' PublicKey (Maybe UTCTime)

The starting time of validity of the public key.

Resource

data Resource

Specifies the type and name of a resource referenced by an event.

See: resource smart constructor.

resource :: Resource

Creates a value of Resource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rResourceType - The type of a resource referenced by the event returned. When the resource type cannot be determined, null is returned. Some examples of resource types are: Instance for EC2, Trail for CloudTrail, DBInstance for RDS, and AccessKey for IAM. For a list of resource types supported for event lookup, see Resource Types Supported for Event Lookup .

rResourceName - The name of the resource referenced by the event returned. These are user-created names whose values will depend on the environment. For example, the resource name might be "auto-scaling-test-group" for an Auto Scaling Group or "i-1234567" for an EC2 Instance.

rResourceType :: Lens' Resource (Maybe Text)

The type of a resource referenced by the event returned. When the resource type cannot be determined, null is returned. Some examples of resource types are: Instance for EC2, Trail for CloudTrail, DBInstance for RDS, and AccessKey for IAM. For a list of resource types supported for event lookup, see Resource Types Supported for Event Lookup .

rResourceName :: Lens' Resource (Maybe Text)

The name of the resource referenced by the event returned. These are user-created names whose values will depend on the environment. For example, the resource name might be "auto-scaling-test-group" for an Auto Scaling Group or "i-1234567" for an EC2 Instance.

ResourceTag

data ResourceTag

A resource tag.

See: resourceTag smart constructor.

resourceTag :: ResourceTag

Creates a value of ResourceTag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rResourceId - Specifies the ARN of the resource.

rTagsList - A list of tags.

rResourceId :: Lens' ResourceTag (Maybe Text)

Specifies the ARN of the resource.

rTagsList :: Lens' ResourceTag [Tag]

A list of tags.

Tag

data Tag

A custom key-value pair associated with a resource such as a CloudTrail trail.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - The value in a key-value pair of a tag. The value must be no longer than 256 Unicode characters.

tagKey - The key in a key-value pair. The key must be must be no longer than 128 Unicode characters. The key must be unique for the resource to which it applies.

tagValue :: Lens' Tag (Maybe Text)

The value in a key-value pair of a tag. The value must be no longer than 256 Unicode characters.

tagKey :: Lens' Tag Text

The key in a key-value pair. The key must be must be no longer than 128 Unicode characters. The key must be unique for the resource to which it applies.

Trail

data Trail

The settings for a trail.

See: trail smart constructor.

trail :: Trail

Creates a value of Trail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tLogFileValidationEnabled - Specifies whether log file validation is enabled.

tTrailARN - Specifies the ARN of the trail. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

tS3KeyPrefix - Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files .The maximum length is 200 characters.

tSNSTopicARN - Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered. The format of a topic ARN is: arn:aws:sns:us-east-1:123456789012:MyTopic

tSNSTopicName - This field is deprecated. Use SnsTopicARN.

tCloudWatchLogsLogGroupARN - Specifies an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered.

tKMSKeyId - Specifies the KMS key ID that encrypts the logs delivered by CloudTrail. The value is a fully specified ARN to a KMS key in the format: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

tHomeRegion - The region in which the trail was created.

tName - Name of the trail set by calling CreateTrail . The maximum length is 128 characters.

tIncludeGlobalServiceEvents - Set to True to include AWS API calls from AWS global services such as IAM. Otherwise, False .

tHasCustomEventSelectors - Specifies if the trail has custom event selectors.

tCloudWatchLogsRoleARN - Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

tS3BucketName - Name of the Amazon S3 bucket into which CloudTrail delivers your trail files. See Amazon S3 Bucket Naming Requirements .

tIsMultiRegionTrail - Specifies whether the trail belongs only to one region or exists in all regions.

tLogFileValidationEnabled :: Lens' Trail (Maybe Bool)

Specifies whether log file validation is enabled.

tTrailARN :: Lens' Trail (Maybe Text)

Specifies the ARN of the trail. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

tS3KeyPrefix :: Lens' Trail (Maybe Text)

Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files .The maximum length is 200 characters.

tSNSTopicARN :: Lens' Trail (Maybe Text)

Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered. The format of a topic ARN is: arn:aws:sns:us-east-1:123456789012:MyTopic

tSNSTopicName :: Lens' Trail (Maybe Text)

This field is deprecated. Use SnsTopicARN.

tCloudWatchLogsLogGroupARN :: Lens' Trail (Maybe Text)

Specifies an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered.

tKMSKeyId :: Lens' Trail (Maybe Text)

Specifies the KMS key ID that encrypts the logs delivered by CloudTrail. The value is a fully specified ARN to a KMS key in the format: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

tHomeRegion :: Lens' Trail (Maybe Text)

The region in which the trail was created.

tName :: Lens' Trail (Maybe Text)

Name of the trail set by calling CreateTrail . The maximum length is 128 characters.

tIncludeGlobalServiceEvents :: Lens' Trail (Maybe Bool)

Set to True to include AWS API calls from AWS global services such as IAM. Otherwise, False .

tHasCustomEventSelectors :: Lens' Trail (Maybe Bool)

Specifies if the trail has custom event selectors.

tCloudWatchLogsRoleARN :: Lens' Trail (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

tS3BucketName :: Lens' Trail (Maybe Text)

Name of the Amazon S3 bucket into which CloudTrail delivers your trail files. See Amazon S3 Bucket Naming Requirements .

tIsMultiRegionTrail :: Lens' Trail (Maybe Bool)

Specifies whether the trail belongs only to one region or exists in all regions.