Network.AWS.CloudTrail.Types

Service

data CloudTrail

Version 2013-11-01 of the Amazon CloudTrail service.

Error

data JSONError :: *

Trail

data Trail

trail :: Trail

Trail constructor.

The fields accessible through corresponding lenses are:

tCloudWatchLogsLogGroupArn :: Maybe Text

tCloudWatchLogsRoleArn :: Maybe Text

tIncludeGlobalServiceEvents :: Maybe Bool

tName :: Maybe Text

tS3BucketName :: Maybe Text

tS3KeyPrefix :: Maybe Text

tSnsTopicName :: Maybe Text

tCloudWatchLogsLogGroupArn :: Lens' Trail (Maybe Text)

Specifies an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered.

tCloudWatchLogsRoleArn :: Lens' Trail (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user’s log group.

tIncludeGlobalServiceEvents :: Lens' Trail (Maybe Bool)

Set to True to include AWS API calls from AWS global services such as IAM. Otherwise, False.

tName :: Lens' Trail (Maybe Text)

Name of the trail set by calling CreateTrail.

tS3BucketName :: Lens' Trail (Maybe Text)

Name of the Amazon S3 bucket into which CloudTrail delivers your trail files.

tS3KeyPrefix :: Lens' Trail (Maybe Text)

Value of the Amazon S3 prefix.

tSnsTopicName :: Lens' Trail (Maybe Text)

Name of the existing Amazon SNS topic that CloudTrail uses to notify the account owner when new CloudTrail log files have been delivered.