Network.AWS.CloudTrail.Types

Service

data CloudTrail

Version 2013-11-01 of the Amazon CloudTrail service.

Error

data JSONError :: *

Event

data Event

event :: Event

Event constructor.

The fields accessible through corresponding lenses are:

eCloudTrailEvent :: Maybe Text

eEventId :: Maybe Text

eEventName :: Maybe Text

eEventTime :: Maybe UTCTime

eResources :: [Resource]

eUsername :: Maybe Text

eCloudTrailEvent :: Lens' Event (Maybe Text)

A JSON string that contains a representation of the event returned.

eEventId :: Lens' Event (Maybe Text)

The CloudTrail ID of the event returned.

eEventName :: Lens' Event (Maybe Text)

The name of the event returned.

eEventTime :: Lens' Event (Maybe UTCTime)

The date and time of the event returned.

eResources :: Lens' Event [Resource]

A list of resources referenced by the event returned.

eUsername :: Lens' Event (Maybe Text)

A user name or role name of the requester that called the API in the event returned.

Trail

data Trail

trail :: Trail

Trail constructor.

The fields accessible through corresponding lenses are:

tCloudWatchLogsLogGroupArn :: Maybe Text

tCloudWatchLogsRoleArn :: Maybe Text

tIncludeGlobalServiceEvents :: Maybe Bool

tName :: Maybe Text

tS3BucketName :: Maybe Text

tS3KeyPrefix :: Maybe Text

tSnsTopicName :: Maybe Text

tCloudWatchLogsLogGroupArn :: Lens' Trail (Maybe Text)

Specifies an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered.

tCloudWatchLogsRoleArn :: Lens' Trail (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user’s log group.

tIncludeGlobalServiceEvents :: Lens' Trail (Maybe Bool)

Set to True to include AWS API calls from AWS global services such as IAM. Otherwise, False.

tName :: Lens' Trail (Maybe Text)

Name of the trail set by calling CreateTrail.

tS3BucketName :: Lens' Trail (Maybe Text)

Name of the Amazon S3 bucket into which CloudTrail delivers your trail files.

tS3KeyPrefix :: Lens' Trail (Maybe Text)

Value of the Amazon S3 prefix.

tSnsTopicName :: Lens' Trail (Maybe Text)

Name of the existing Amazon SNS topic that CloudTrail uses to notify the account owner when new CloudTrail log files have been delivered.

LookupAttribute

data LookupAttribute

lookupAttribute

Arguments

:: LookupAttributeKey
laAttributeKey

-> Text
laAttributeValue

-> LookupAttribute

LookupAttribute constructor.

The fields accessible through corresponding lenses are:

laAttributeKey :: LookupAttributeKey

laAttributeValue :: Text

laAttributeKey :: Lens' LookupAttribute LookupAttributeKey

Specifies an attribute on which to filter the events returned.

laAttributeValue :: Lens' LookupAttribute Text

Specifies a value for the specified AttributeKey.

LookupAttributeKey

data LookupAttributeKey

Constructors

EventId
EventId

EventName
EventName

ResourceName
ResourceName

ResourceType
ResourceType

Username
Username

Resource

data Resource

resource :: Resource

Resource constructor.

The fields accessible through corresponding lenses are:

rResourceName :: Maybe Text

rResourceType :: Maybe Text

rResourceName :: Lens' Resource (Maybe Text)

The name of the resource referenced by the event returned. These are user-created names whose values will depend on the environment. For example, the resource name might be "auto-scaling-test-group" for an Auto Scaling Group or "i-1234567" for an EC2 Instance.

rResourceType :: Lens' Resource (Maybe Text)

The type of a resource referenced by the event returned. When the resource type cannot be determined, null is returned. Some examples of resource types are: Instance for EC2, Trail for CloudTrail, DBInstance for RDS, and AccessKey for IAM. For a list of resource types supported for event lookup, see Resource Types Supported for Event Lookup.