Network.AWS.CloudTrail.UpdateTrail

Description

Updates the settings that specify delivery of log files. Changes to a trail do not require stopping the CloudTrail service. Use this action to designate an existing bucket for log delivery. If the existing bucket has previously been a target for CloudTrail log files, an IAM policy exists for the bucket. UpdateTrail must be called from the region in which the trail was created; otherwise, an InvalidHomeRegionException is thrown.

Creating a Request

updateTrail

Arguments

:: Text
utName

-> UpdateTrail

Creates a value of UpdateTrail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utS3KeyPrefix - Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files . The maximum length is 200 characters.

utSNSTopicName - Specifies the name of the Amazon SNS topic defined for notification of log file delivery. The maximum length is 256 characters.

utEnableLogFileValidation - Specifies whether log file validation is enabled. The default is false.

utCloudWatchLogsLogGroupARN - Specifies a log group name using an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered. Not required unless you specify CloudWatchLogsRoleArn.

utKMSKeyId - Specifies the KMS key ID to use to encrypt the logs delivered by CloudTrail. The value can be an alias name prefixed by "alias", a fully specified ARN to an alias, a fully specified ARN to a key, or a globally unique identifier. Examples: * aliasMyAliasName * arn:aws:kms:us-east-1:123456789012:aliasMyAliasName * arn:aws:kms:us-east-1:123456789012:key12345678-1234-1234-1234-123456789012 * 12345678-1234-1234-1234-123456789012

utIncludeGlobalServiceEvents - Specifies whether the trail is publishing events from global services such as IAM to the log files.

utCloudWatchLogsRoleARN - Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

utS3BucketName - Specifies the name of the Amazon S3 bucket designated for publishing log files. See Amazon S3 Bucket Naming Requirements .

utIsMultiRegionTrail - Specifies whether the trail applies only to the current region or to all regions. The default is false. If the trail exists only in the current region and this value is set to true, shadow trails (replications of the trail) will be created in the other regions. If the trail exists in all regions and this value is set to false, the trail will remain in the region where it was created, and its shadow trails in other regions will be deleted.

utName - Specifies the name of the trail or trail ARN. If Name is a trail name, the string must meet the following requirements: * Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-) * Start with a letter or number, and end with a letter or number * Be between 3 and 128 characters * Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are invalid. * Not be in IP address format (for example, 192.168.5.4) If Name is a trail ARN, it must be in the format: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

data UpdateTrail

Specifies settings to update for the trail.

See: updateTrail smart constructor.

Request Lenses

utS3KeyPrefix :: Lens' UpdateTrail (Maybe Text)

Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files . The maximum length is 200 characters.

utSNSTopicName :: Lens' UpdateTrail (Maybe Text)

Specifies the name of the Amazon SNS topic defined for notification of log file delivery. The maximum length is 256 characters.

utEnableLogFileValidation :: Lens' UpdateTrail (Maybe Bool)

Specifies whether log file validation is enabled. The default is false.

utCloudWatchLogsLogGroupARN :: Lens' UpdateTrail (Maybe Text)

Specifies a log group name using an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered. Not required unless you specify CloudWatchLogsRoleArn.

utKMSKeyId :: Lens' UpdateTrail (Maybe Text)

Specifies the KMS key ID to use to encrypt the logs delivered by CloudTrail. The value can be an alias name prefixed by "alias", a fully specified ARN to an alias, a fully specified ARN to a key, or a globally unique identifier. Examples: * aliasMyAliasName * arn:aws:kms:us-east-1:123456789012:aliasMyAliasName * arn:aws:kms:us-east-1:123456789012:key12345678-1234-1234-1234-123456789012 * 12345678-1234-1234-1234-123456789012

utIncludeGlobalServiceEvents :: Lens' UpdateTrail (Maybe Bool)

Specifies whether the trail is publishing events from global services such as IAM to the log files.

utCloudWatchLogsRoleARN :: Lens' UpdateTrail (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

utS3BucketName :: Lens' UpdateTrail (Maybe Text)

Specifies the name of the Amazon S3 bucket designated for publishing log files. See Amazon S3 Bucket Naming Requirements .

utIsMultiRegionTrail :: Lens' UpdateTrail (Maybe Bool)

Specifies whether the trail applies only to the current region or to all regions. The default is false. If the trail exists only in the current region and this value is set to true, shadow trails (replications of the trail) will be created in the other regions. If the trail exists in all regions and this value is set to false, the trail will remain in the region where it was created, and its shadow trails in other regions will be deleted.

utName :: Lens' UpdateTrail Text

Specifies the name of the trail or trail ARN. If Name is a trail name, the string must meet the following requirements: * Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-) * Start with a letter or number, and end with a letter or number * Be between 3 and 128 characters * Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are invalid. * Not be in IP address format (for example, 192.168.5.4) If Name is a trail ARN, it must be in the format: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

Destructuring the Response

updateTrailResponse

Arguments

:: Int
utrsResponseStatus

-> UpdateTrailResponse

Creates a value of UpdateTrailResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utrsLogFileValidationEnabled - Specifies whether log file integrity validation is enabled.

utrsTrailARN - Specifies the ARN of the trail that was updated. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

utrsS3KeyPrefix - Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files .

utrsSNSTopicARN - Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered. The format of a topic ARN is: arn:aws:sns:us-east-1:123456789012:MyTopic

utrsSNSTopicName - This field is deprecated. Use SnsTopicARN.

utrsCloudWatchLogsLogGroupARN - Specifies the Amazon Resource Name (ARN) of the log group to which CloudTrail logs will be delivered.

utrsKMSKeyId - Specifies the KMS key ID that encrypts the logs delivered by CloudTrail. The value is a fully specified ARN to a KMS key in the format: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

utrsName - Specifies the name of the trail.

utrsIncludeGlobalServiceEvents - Specifies whether the trail is publishing events from global services such as IAM to the log files.

utrsCloudWatchLogsRoleARN - Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

utrsS3BucketName - Specifies the name of the Amazon S3 bucket designated for publishing log files.

utrsIsMultiRegionTrail - Specifies whether the trail exists in one region or in all regions.

utrsResponseStatus - -- | The response status code.

data UpdateTrailResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: updateTrailResponse smart constructor.

Response Lenses

utrsLogFileValidationEnabled :: Lens' UpdateTrailResponse (Maybe Bool)

Specifies whether log file integrity validation is enabled.

utrsTrailARN :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the ARN of the trail that was updated. The format of a trail ARN is: arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

utrsS3KeyPrefix :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files .

utrsSNSTopicARN :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered. The format of a topic ARN is: arn:aws:sns:us-east-1:123456789012:MyTopic

utrsSNSTopicName :: Lens' UpdateTrailResponse (Maybe Text)

This field is deprecated. Use SnsTopicARN.

utrsCloudWatchLogsLogGroupARN :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the Amazon Resource Name (ARN) of the log group to which CloudTrail logs will be delivered.

utrsKMSKeyId :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the KMS key ID that encrypts the logs delivered by CloudTrail. The value is a fully specified ARN to a KMS key in the format: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

utrsName :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the name of the trail.

utrsIncludeGlobalServiceEvents :: Lens' UpdateTrailResponse (Maybe Bool)

Specifies whether the trail is publishing events from global services such as IAM to the log files.

utrsCloudWatchLogsRoleARN :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

utrsS3BucketName :: Lens' UpdateTrailResponse (Maybe Text)

Specifies the name of the Amazon S3 bucket designated for publishing log files.

utrsIsMultiRegionTrail :: Lens' UpdateTrailResponse (Maybe Bool)

Specifies whether the trail exists in one region or in all regions.

utrsResponseStatus :: Lens' UpdateTrailResponse Int

- | The response status code.