Network.AWS.CloudTrail.V2013_11_01

Description

CloudTrail is a web service that records AWS API calls for your AWS account and delivers log files to an Amazon S3 bucket. The recorded information includes the identity of the user, the start time of the AWS API call, the source IP address, the request parameters, and the response elements returned by the service.

Documentation

createTrail

Arguments

:: Text
_ctrName

-> Text
_ctrS3BucketName

-> CreateTrail

Minimum specification for a CreateTrail request.

data CreateTrail

Constructors

CreateTrail

Fields

_ctrName :: Text

Specifies the name of the trail.

_ctrS3BucketName :: Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

_ctrIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_ctrS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_ctrSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

ctrSnsTopicName :: Lens' CreateTrail (Maybe Text)

ctrS3KeyPrefix :: Lens' CreateTrail (Maybe Text)

ctrS3BucketName :: Lens' CreateTrail Text

ctrName :: Lens' CreateTrail Text

ctrIncludeGlobalServiceEvents :: Lens' CreateTrail (Maybe Bool)

data CreateTrailResponse

Constructors

CreateTrailResponse

Fields

_ctsIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_ctsS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_ctsSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

_ctsName :: Maybe Text

Specifies the name of the trail.

_ctsS3BucketName :: Maybe Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

ctsSnsTopicName :: Lens' CreateTrailResponse (Maybe Text)

ctsS3KeyPrefix :: Lens' CreateTrailResponse (Maybe Text)

ctsS3BucketName :: Lens' CreateTrailResponse (Maybe Text)

ctsName :: Lens' CreateTrailResponse (Maybe Text)

ctsIncludeGlobalServiceEvents :: Lens' CreateTrailResponse (Maybe Bool)

data DeleteTrail

Constructors

DeleteTrail

Fields

_dttName :: Text

The name of a trail to be deleted.

dttName :: Iso' DeleteTrail Text

data DeleteTrailResponse

Constructors

DeleteTrailResponse

describeTrails :: DescribeTrails

Minimum specification for a DescribeTrails request.

data DescribeTrails

Constructors

DescribeTrails

Fields

_dtrTrailNameList :: [Text]

The trail returned.

dtrTrailNameList :: Iso' DescribeTrails [Text]

data DescribeTrailsResponse

Constructors

DescribeTrailsResponse

Fields

_dtsTrailList :: [Trail]

The list of trails.

dtsTrailList :: Iso' DescribeTrailsResponse [Trail]

data GetTrailStatus

Constructors

GetTrailStatus

Fields

_gtsrName :: Text

The name of the trail for which you are requesting the current status.

gtsrName :: Iso' GetTrailStatus Text

data GetTrailStatusResponse

Constructors

GetTrailStatusResponse

Fields

_gtssIsLogging :: Maybe Bool

Whether the CloudTrail is currently logging AWS API calls.

_gtssStartLoggingTime :: Maybe ISO8601

Specifies the most recent date and time when CloudTrail started recording API calls for an AWS account.

_gtssLatestDeliveryTime :: Maybe ISO8601

Specifies the date and time that CloudTrail last delivered log files to an account's Amazon S3 bucket.

_gtssLatestNotificationTime :: Maybe ISO8601

Specifies the date and time of the most recent Amazon SNS notification that CloudTrail has written a new log file to an account's Amazon S3 bucket.

_gtssStopLoggingTime :: Maybe ISO8601

Specifies the most recent date and time when CloudTrail stopped recording API calls for an AWS account.

_gtssLatestDeliveryError :: Maybe Text

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver log files to the designated bucket. For more information see the topic Error Responses in the Amazon S3 API Reference.

_gtssLatestNotificationError :: Maybe Text

Displays any Amazon SNS error that CloudTrail encountered when attempting to send a notification. For more information about Amazon SNS errors, see the Amazon SNS Developer Guide.

gtssStopLoggingTime :: Lens' GetTrailStatusResponse (Maybe ISO8601)

gtssStartLoggingTime :: Lens' GetTrailStatusResponse (Maybe ISO8601)

gtssLatestNotificationTime :: Lens' GetTrailStatusResponse (Maybe ISO8601)

gtssLatestNotificationError :: Lens' GetTrailStatusResponse (Maybe Text)

gtssLatestDeliveryTime :: Lens' GetTrailStatusResponse (Maybe ISO8601)

gtssLatestDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

gtssIsLogging :: Lens' GetTrailStatusResponse (Maybe Bool)

data StartLogging

Constructors

StartLogging

Fields

_sltName :: Text

The name of the trail for which CloudTrail logs AWS API calls.

sltName :: Iso' StartLogging Text

data StartLoggingResponse

Constructors

StartLoggingResponse

data StopLogging

Constructors

StopLogging

Fields

_slrName :: Text

Communicates to CloudTrail the name of the trail for which to stop logging AWS API calls.

slrName :: Iso' StopLogging Text

data StopLoggingResponse

Constructors

StopLoggingResponse

data CloudTrail

Supported version (2013-11-01) of the AWS CloudTrail service.

data Trail

The settings for a trail.

Constructors

Trail

Fields

_tS3KeyPrefix :: Maybe Text

Value of the Amazon S3 prefix.

_tSnsTopicName :: Maybe Text

Name of the existing Amazon SNS topic that CloudTrail uses to notify the account owner when new CloudTrail log files have been delivered.

_tName :: Maybe Text

Name of the trail set by calling CreateTrail.

_tIncludeGlobalServiceEvents :: Maybe Bool

Set to True to include AWS API calls from AWS global services such as IAM. Otherwise, False.

_tS3BucketName :: Maybe Text

Name of the Amazon S3 bucket into which CloudTrail delivers your trail files.

tSnsTopicName :: Lens' Trail (Maybe Text)

tS3KeyPrefix :: Lens' Trail (Maybe Text)

tS3BucketName :: Lens' Trail (Maybe Text)

tName :: Lens' Trail (Maybe Text)

tIncludeGlobalServiceEvents :: Lens' Trail (Maybe Bool)

updateTrail

Arguments

:: Text
_utrName

-> UpdateTrail

Minimum specification for a UpdateTrail request.

data UpdateTrail

Constructors

UpdateTrail

Fields

_utrName :: Text

Specifies the name of the trail.

_utrIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_utrS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_utrSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

_utrS3BucketName :: Maybe Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

utrSnsTopicName :: Lens' UpdateTrail (Maybe Text)

utrS3KeyPrefix :: Lens' UpdateTrail (Maybe Text)

utrS3BucketName :: Lens' UpdateTrail (Maybe Text)

utrName :: Lens' UpdateTrail Text

utrIncludeGlobalServiceEvents :: Lens' UpdateTrail (Maybe Bool)

data UpdateTrailResponse

Constructors

UpdateTrailResponse

Fields

_utsIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_utsS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_utsSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

_utsName :: Maybe Text

Specifies the name of the trail.

_utsS3BucketName :: Maybe Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

utsSnsTopicName :: Lens' UpdateTrailResponse (Maybe Text)

utsS3KeyPrefix :: Lens' UpdateTrailResponse (Maybe Text)

utsS3BucketName :: Lens' UpdateTrailResponse (Maybe Text)

utsName :: Lens' UpdateTrailResponse (Maybe Text)

utsIncludeGlobalServiceEvents :: Lens' UpdateTrailResponse (Maybe Bool)